refactor: LadderGameMachine 패키지 이동
