+++ src/main/java/ladder/Application.java
@@ -1,7 +1,5 @@
 package ladder;
 
-import ladder.controller.LadderGameMachine;
-
 public class Application {
     public static void main(String[] args) {
         LadderGameMachine gameMachine = new LadderGameMachine();
